--- internal/component/database_observability/collector/types.go
@@ -1,8 +1,0 @@
-package collector
-
-import "context"
-
-type Collector interface {
-	Run(context.Context) error
-	Stop()
-}
