+++ internal/component/database_observability/mysql/collector/connection_info.go
